+++ lib/templates/middleAssignment/main.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'home_page.dart';
+
+void main() {
+  runApp(ProviderScope(child: MiddleAssigmentApp()));
+}
+
+class MiddleAssigmentApp extends StatelessWidget {
+  const MiddleAssigmentApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      title: 'Middle Assigment - contrast shower',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: ContrastShowerScreenHome(),
+    );
+  }
+}
